Backend updates
I installed express.js to connect the backend to the frontend, I did some query testing to see if I can retrieve the data from the database in the backend. That is confirmed. Now the next steps is to connect the the usernames and passwords from the database to the login page, and then we can finish it all up with a log out button that will log the daily progress of each user into the daily progress table.